fix: add individual data save
This commit refactors the graph saving logic in the `save_set_graph` function. It removes the commented out code for saving individual graphs and instead adds a new section to save individual data plots. If the `save_individual` flag is set to true and the data is time series, it generates a separate graph for each individual data and saves it with the original output name appended with "_individual". This improves the organization and clarity of the code.

Co-authored-by: sakashita44